Add the join mailing list page
