--- main.py
@@ -1,17 +1,0 @@
-from flask import Flask,render_template
-
-app = Flask(__name__)
-
-@app.route('/')
-def home():
-    return render_template('home.html')
-
-@app.route('/about')
-def about():
-    return render_template('about.html')
-
-@app.route("/learn")
-def learn():
-    return render_template("learn.html")
-if __name__=='__main__':
-    app.run(debug=True)
