+++ templates/join.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>PyKanpur</title>
+</head>
+<>
+<h2>Join the mailing list</h2><br>
+<form method="post">
+    E-Mail:<input type="text" name="mail">
+    <button type = "submit">Join Now!</button>
+</form>
+</body>
+</html>
