+++ templates/success.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>PyKanpur</title>
+</head>
+<body>
+<ul>
+{%for i in mails%}
+    <li>{%print(i)%}
+{%endfor%}
+</ul>
+</body>
+</html>
